requirements.txt
pyaudio
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
torch
git+https://github.com/openai/whisper.git
soundfile
sounddevice

llama-cpp-python
librosa
unidecode
inflect
matplotlib
wave

# Python version 3.8 works just fine, regarding librosa, the newest version renders bugs in Real-Time-Voice-Cloning, but it takes repairing only 2 basic functions to repair that
#RTVC requirements:
numpy==1.20.3
Pillow==8.4.0
PyQt5==5.15.6
scikit-learn==1.0.2
scipy==1.7.3
sounddevice==0.4.3
SoundFile==0.10.3.post1
tqdm==4.62.3
umap-learn==0.5.2
urllib3==1.26.7
visdom==0.1.8.9
webrtcvad==2.0.10